Streaming processing of DIADEM test datasets.
Before all the training and validation instances were loaded into memory
by DIADEM.
Now, the validation dataset can be processed as a stream, which allows
to validate detection models on bigger datasets.
Stream processing is available only with the option
--validation-mode ValidationDataset --validation-dataset <validation_dataset>.
In this case, the alerts cannot be analyzed with clustering or
classification (since these analyses require to load all the data into
memory).