[NEW FEATURE] Features types can be specified.
When the types are specified (in the file <features>_description.csv),
SecuML does not have to induce them automatically when the dataset is
loaded for the first time. It reduces the execution time.

For now, two types are supported: numeric and binary.

The datasets SpamHam/lingspam, and Gaussian/test have been updated to
show examples of files <features>_description.csv.