New version of scikit-learn (0.21.1).
The new version of scikit-learn is not compatible with the current
version of metri-learn available on PyPI. A new version should be
released shortly
(https://github.com/metric-learn/metric-learn/issues/209).

In the mean time, SecuML depends on an unreleased version of
metric-learn
(git+https://github.com/ab-anssi/metric-learn.git@v0.4.0ab#egg=metric-learn-0.4.0ab),
that is compatible with scikit-learn 0.21.1.